Spring cleaning!
Replaces uses of removed/deprecated API:s with their non-deprecated
counterparts, and fixes a number of new NP-warnings.

+++ backend/core/src/main/java/com/elvaco/mvp/core/spi/data/Pageable.java
@@ -5,7 +5,7 @@ public interface Pageable {
 
   int getPageSize();
 
-  int getOffset();
+  long getOffset();
 
   Sort getSort();
